Add sidebar to compaction docs page
This ticket adds the sidebar to the Compaction docs page and refactors
the code a little to start explictly stating what resources and Views
in the documents addon are intended for sharing.

Closes COUCHDB-2404
