+++ app/addons/documents/shared-views.js
